fix: decode client_id as u64

+++ yrs/src/updates/decoder.rs
@@ -83,7 +83,7 @@ impl<'a> DecoderV1<'a> {
     }
 
     fn read_id(&mut self) -> Result<ID, Error> {
-        let client: u32 = self.read_var()?;
+        let client: u64 = self.read_var()?;
         let clock = self.read_var()?;
         Ok(ID::new(client as ClientID, clock))
     }
@@ -141,7 +141,7 @@ impl<'a> Decoder for DecoderV1<'a> {
 
     #[inline]
     fn read_client(&mut self) -> Result<ClientID, Error> {
-        let client: u32 = self.cursor.read_var()?;
+        let client: u64 = self.cursor.read_var()?;
         Ok(client as ClientID)
     }
